Move lazy-helpers to separate file
