+++ backend/tilavarauspalvelu/models/allocated_timeslot/model.py
@@ -10,7 +10,7 @@ from lookup_property import L, lookup_property
 
 from tilavarauspalvelu.enums import Weekday
 from utils.fields.model import StrChoiceField
-from utils.utils import LazyModelAttribute, LazyModelManager
+from utils.lazy import LazyModelAttribute, LazyModelManager
 
 if TYPE_CHECKING:
     import datetime
