+++ backend/tilavarauspalvelu/models/application_round/model.py
@@ -15,7 +15,7 @@ from tilavarauspalvelu.enums import (
 )
 from utils.date_utils import local_datetime
 from utils.db import NowTT
-from utils.utils import LazyModelAttribute, LazyModelManager
+from utils.lazy import LazyModelAttribute, LazyModelManager
 
 if TYPE_CHECKING:
     from tilavarauspalvelu.models import TermsOfUse, Unit
