+++ backend/tilavarauspalvelu/models/location/model.py
@@ -7,7 +7,7 @@ from django.db import models
 from django.utils.translation import gettext_lazy as _
 
 from tilavarauspalvelu.constants import COORDINATE_SYSTEM_ID
-from utils.utils import LazyModelAttribute, LazyModelManager
+from utils.lazy import LazyModelAttribute, LazyModelManager
 
 if TYPE_CHECKING:
     from django.contrib.gis.geos import Point
