+++ backend/tilavarauspalvelu/models/reservable_time_span/model.py
@@ -7,7 +7,7 @@ from django.utils.translation import gettext_lazy as _
 
 from tilavarauspalvelu.integrations.opening_hours.time_span_element import TimeSpanElement
 from utils.date_utils import DEFAULT_TIMEZONE
-from utils.utils import LazyModelAttribute, LazyModelManager
+from utils.lazy import LazyModelAttribute, LazyModelManager
 
 if TYPE_CHECKING:
     import datetime
