+++ backend/tilavarauspalvelu/models/reservation_statistic_unit/model.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING, ClassVar
 from django.db import models, transaction
 from django.utils.translation import gettext_lazy as _
 
-from utils.utils import LazyModelAttribute, LazyModelManager
+from utils.lazy import LazyModelAttribute, LazyModelManager
 
 if TYPE_CHECKING:
     from tilavarauspalvelu.models import ReservationStatistic
