+++ backend/tilavarauspalvelu/models/reservation_unit_cancellation_rule/model.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING, ClassVar
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
-from utils.utils import LazyModelAttribute, LazyModelManager
+from utils.lazy import LazyModelAttribute, LazyModelManager
 
 if TYPE_CHECKING:
     from .actions import ReservationUnitCancellationRuleActions
