+++ backend/tilavarauspalvelu/models/reservation_unit_image/model.py
@@ -8,7 +8,7 @@ from django.utils.translation import gettext_lazy as _
 from easy_thumbnails.fields import ThumbnailerImageField
 
 from tilavarauspalvelu.enums import ReservationUnitImageType
-from utils.utils import LazyModelAttribute, LazyModelManager
+from utils.lazy import LazyModelAttribute, LazyModelManager
 
 if TYPE_CHECKING:
     from easy_thumbnails.files import ThumbnailerImageFieldFile
