+++ backend/tilavarauspalvelu/models/sql_log/model.py
@@ -7,7 +7,7 @@ import sqlparse
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
-from utils.utils import LazyModelAttribute, LazyModelManager
+from utils.lazy import LazyModelAttribute, LazyModelManager
 
 if TYPE_CHECKING:
     from .actions import SQLLogActions
